move fixtures into their own file
