+++ src/export/export_accounting.rs
@@ -93,7 +93,8 @@ mod tests {
     use pretty_assertions::assert_eq;
     use rstest::rstest;
 
-    use crate::test_utils::{assert_dataframe, sample_date_01, summary_df_01};
+    use crate::test_fixtures::{sample_date_01, summary_df_01};
+    use crate::test_utils::assert_dataframe;
 
     use super::*;
